Conflicts:
	Core/Code/Testing/mitkGeometry3DEqualTest.cpp
	Modules/ContourModel/DataManagement/mitkContourModel.cpp
	Modules/IGT/IGTFilters/mitkNavigationDataObjectVisualizationFilter.cpp

+++ Core/Code/Controllers/mitkSliceNavigationController.cpp
@@ -575,7 +575,7 @@ SliceNavigationController::AdjustSliceStepperRange()
   int i, k = 0;
   for ( i = 0; i < 3; ++i )
   {
-    if ( fabs( (float) direction[i] ) < 0.000000001 ) { ++c; }
+    if ( fabs(direction[i]) < 0.000000001 ) { ++c; }
     else { k = i; }
   }
